search updated everytime
